BE/requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements-dev.in
#
appdirs==1.4.4
    # via black
atomicwrites==1.4.0
    # via pytest
attrs==20.3.0
    # via pytest
black==20.8b1
    # via -r requirements-dev.in
certifi==2021.5.30
    # via sentry-sdk
click==7.1.2
    # via
    #   black
    #   face-recognition
    #   uvicorn
colorama==0.4.4
    # via pytest
databases==0.4.3
    # via -r requirements-build.in
dlib==19.21.1
    # via face-recognition
face-recognition==1.3.0
    # via -r requirements-build.in
face-recognition-models==0.3.0
    # via face-recognition
h11==0.12.0
    # via uvicorn
iniconfig==1.1.1
    # via pytest
isort==5.7.0
    # via -r requirements-dev.in
marshmallow==3.13.0
    # via -r requirements-build.in
mongoengine==0.22.1
    # via -r requirements-build.in
mypy==0.910
    # via -r requirements-dev.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
numpy==1.20.1
    # via face-recognition
packaging==20.9
    # via pytest
pathspec==0.8.1
    # via black
pillow==8.1.0
    # via
    #   -r requirements-build.in
    #   face-recognition
pluggy==0.13.1
    # via pytest
py==1.10.0
    # via pytest
pymongo==3.11.3
    # via mongoengine
pyparsing==2.4.7
    # via packaging
pytest==6.2.2
    # via -r requirements-dev.in
python-multipart==0.0.5
    # via -r requirements-build.in
regex==2020.11.13
    # via black
sentry-sdk==1.3.0
    # via -r requirements-build.in
six==1.16.0
    # via python-multipart
sqlalchemy==1.3.24
    # via databases
starlette==0.14.2
    # via -r requirements-build.in
toml==0.10.2
    # via
    #   black
    #   mypy
    #   pytest
typed-ast==1.4.2
    # via black
typing-extensions==3.7.4.3
    # via
    #   black
    #   mypy
urllib3==1.26.6
    # via sentry-sdk
uvicorn==0.13.4
    # via -r requirements-build.in
